test: Group utility for random byte slice generation
